Convert over all clan commands
