--- src/commands/clan/index.ts
@@ -1,5 +1,0 @@
-import brains from "./brains";
-import skills from "./skills";
-import status from "./status";
-
-export default { brains, skills, status };
