+++ src/db.ts
@@ -1,0 +1,6 @@
+import { Pool } from "pg";
+
+export const pool = new Pool({
+  connectionString: process.env.DATABASE_URL,
+  ssl: false,
+});
